server/app/requirements.txt
aiofiles==0.5.0
amqp==5.0.6
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.3.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
orjson==3.5.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.18
py==1.10.0
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.16.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
